collector: split backtest collector support owners
