+++ src/collector/backtest_collector.rs
@@ -12,12 +12,14 @@
 //! ploy collect-data --live --symbols BTC,ETH,SOL --output ./data/
 //! ```
 
+mod csv_sink;
+mod pm_collection;
+
 use chrono::{DateTime, Duration, Timelike, Utc};
 use rust_decimal::prelude::*;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use std::fs::{create_dir_all, File, OpenOptions};
-use std::io::{BufWriter, Write};
+use std::fs::File;
+use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use tokio::sync::RwLock;
@@ -27,6 +29,10 @@ use tracing::{debug, error, info, warn};
 use crate::adapters::PolymarketClient;
 use crate::collector::BinanceKlineClient;
 use crate::error::{PloyError, Result};
+
+pub use csv_sink::{KlineCSV, PMPriceCSV};
+pub use pm_collection::ActiveMarket;
+use pm_collection::TrackedMarket;
 
 // ============================================================================
 // Configuration
@@ -68,49 +74,6 @@ impl Default for CollectorConfig {
 // ============================================================================
 // Data Structures
 // ============================================================================
-
-/// K-line record for CSV export
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct KlineCSV {
-    pub timestamp: i64,
-    pub datetime: String,
-    pub symbol: String,
-    pub open: String,
-    pub high: String,
-    pub low: String,
-    pub close: String,
-    pub volume: String,
-    pub trades: u64,
-}
-
-/// PM price record for CSV export
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PMPriceCSV {
-    pub timestamp: i64,
-    pub datetime: String,
-    pub market_id: String,
-    pub condition_id: String,
-    pub symbol: String,
-    pub threshold: String,
-    pub spot_price: String,
-    pub yes_price: String,
-    pub no_price: String,
-    pub yes_bid: String,
-    pub yes_ask: String,
-    pub no_bid: String,
-    pub no_ask: String,
-    pub resolution_time: i64,
-    pub time_remaining_secs: i64,
-    pub outcome: String, // "pending", "YES", "NO"
-}
-
-/// Tracked market for resolution monitoring
-#[derive(Debug, Clone)]
-struct TrackedMarket {
-    market_id: String,
-    condition_id: String,
-    resolution_time: DateTime<Utc>,
-}
 
 // ============================================================================
 // Backtest Data Collector
@@ -157,9 +120,6 @@ impl BacktestCollector {
         info!("Output directory: {:?}", self.config.output_dir);
 
         if self.config.persist_csv {
-            // Legacy CSV sink: optional.
-            create_dir_all(&self.config.output_dir)
-                .map_err(|e| PloyError::Internal(format!("Failed to create output dir: {}", e)))?;
             self.init_csv_files()?;
         } else {
             info!("CSV sink disabled (persist_csv=false)");
@@ -199,34 +159,6 @@ impl BacktestCollector {
                     error!("Resolution checker error: {:?}", e);
                 }
             }
-        }
-
-        Ok(())
-    }
-
-    /// Initialize CSV files with headers
-    fn init_csv_files(&self) -> Result<()> {
-        // K-line CSV
-        let kline_path = self.config.output_dir.join("klines.csv");
-        if !kline_path.exists() {
-            let mut file =
-                File::create(&kline_path).map_err(|e| PloyError::Internal(e.to_string()))?;
-            writeln!(
-                file,
-                "timestamp,datetime,symbol,open,high,low,close,volume,trades"
-            )
-            .map_err(|e| PloyError::Internal(e.to_string()))?;
-            info!("Created {}", kline_path.display());
-        }
-
-        // PM price CSV
-        let pm_path = self.config.output_dir.join("pm_prices.csv");
-        if !pm_path.exists() {
-            let mut file =
-                File::create(&pm_path).map_err(|e| PloyError::Internal(e.to_string()))?;
-            writeln!(file, "timestamp,datetime,market_id,condition_id,symbol,threshold,spot_price,yes_price,no_price,yes_bid,yes_ask,no_bid,no_ask,resolution_time,time_remaining_secs,outcome")
-                .map_err(|e| PloyError::Internal(e.to_string()))?;
-            info!("Created {}", pm_path.display());
         }
 
         Ok(())
@@ -308,164 +240,6 @@ impl BacktestCollector {
         })
     }
 
-    /// Spawn PM price collection task
-    fn spawn_pm_collector(&self, end_time: Option<DateTime<Utc>>) -> tokio::task::JoinHandle<()> {
-        let config = self.config.clone();
-        let pm_client = self.pm_client.clone();
-        let stats = self.stats.clone();
-        let tracked_markets = self.tracked_markets.clone();
-        let kline_client = BinanceKlineClient::new();
-
-        tokio::spawn(async move {
-            let mut ticker = interval(std::time::Duration::from_secs(config.pm_interval_secs));
-
-            loop {
-                ticker.tick().await;
-
-                // Check if we should stop
-                if let Some(end) = end_time {
-                    if Utc::now() >= end {
-                        info!("PM collection duration reached, stopping");
-                        break;
-                    }
-                }
-
-                // Skip if no PM client
-                let Some(ref client) = pm_client else {
-                    debug!("No PM client configured, skipping PM collection");
-                    continue;
-                };
-
-                // Collect PM prices for each symbol
-                for symbol in &config.symbols {
-                    // Get current spot price from Binance
-                    let spot_price = match kline_client.fetch_klines(symbol, "1m", 1).await {
-                        Ok(klines) => klines.last().map(|k| k.close).unwrap_or(Decimal::ZERO),
-                        Err(_) => Decimal::ZERO,
-                    };
-
-                    // Fetch active 15-minute markets for this symbol
-                    match Self::fetch_active_markets(client, symbol).await {
-                        Ok(markets) => {
-                            for market in markets {
-                                let write_ok = if config.persist_csv {
-                                    match Self::append_pm_price(
-                                        &config.output_dir,
-                                        &market,
-                                        spot_price,
-                                    )
-                                    .await
-                                    {
-                                        Ok(_) => true,
-                                        Err(e) => {
-                                            warn!("Failed to append PM price: {}", e);
-                                            false
-                                        }
-                                    }
-                                } else {
-                                    true
-                                };
-
-                                if write_ok {
-                                    let mut s = stats.write().await;
-                                    s.pm_prices_collected += 1;
-                                    s.last_pm_time = Some(Utc::now());
-
-                                    // Track for resolution
-                                    let mut tracked = tracked_markets.write().await;
-                                    tracked.insert(
-                                        market.market_id.clone(),
-                                        TrackedMarket {
-                                            market_id: market.market_id.clone(),
-                                            condition_id: market.condition_id.clone(),
-                                            resolution_time: market.resolution_time,
-                                        },
-                                    );
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            debug!("Failed to fetch markets for {}: {}", symbol, e);
-                        }
-                    }
-
-                    sleep(std::time::Duration::from_millis(200)).await;
-                }
-            }
-        })
-    }
-
-    /// Spawn resolution checking task
-    fn spawn_resolution_checker(
-        &self,
-        end_time: Option<DateTime<Utc>>,
-    ) -> tokio::task::JoinHandle<()> {
-        let config = self.config.clone();
-        let pm_client = self.pm_client.clone();
-        let stats = self.stats.clone();
-        let tracked_markets = self.tracked_markets.clone();
-
-        tokio::spawn(async move {
-            let mut ticker = interval(std::time::Duration::from_secs(60)); // Check every minute
-
-            loop {
-                ticker.tick().await;
-
-                // Check if we should stop
-                if let Some(end) = end_time {
-                    if Utc::now() >= end {
-                        info!("Resolution checker duration reached, stopping");
-                        break;
-                    }
-                }
-
-                let Some(ref client) = pm_client else {
-                    continue;
-                };
-
-                // Check tracked markets for resolution
-                let now = Utc::now();
-                let mut resolved = Vec::new();
-
-                {
-                    let tracked = tracked_markets.read().await;
-                    for (_market_id, market) in tracked.iter() {
-                        // If past resolution time, check outcome
-                        if now > market.resolution_time + Duration::minutes(5) {
-                            resolved.push(market.clone());
-                        }
-                    }
-                }
-
-                // Update resolved markets
-                for market in resolved {
-                    if let Ok(outcome) = Self::check_resolution(client, &market.condition_id).await
-                    {
-                        // Update the CSV with outcome
-                        if let Err(e) =
-                            Self::update_outcome(&config.output_dir, &market.market_id, outcome)
-                                .await
-                        {
-                            warn!("Failed to update outcome for {}: {}", market.market_id, e);
-                        } else {
-                            let mut s = stats.write().await;
-                            s.markets_resolved += 1;
-                            info!(
-                                "Market {} resolved: {}",
-                                market.market_id,
-                                if outcome { "YES" } else { "NO" }
-                            );
-                        }
-
-                        // Remove from tracked
-                        let mut tracked = tracked_markets.write().await;
-                        tracked.remove(&market.market_id);
-                    }
-                }
-            }
-        })
-    }
-
     /// Get next 15-minute boundary time
     fn next_15min_boundary(from: DateTime<Utc>) -> DateTime<Utc> {
         let minute = from.minute();
@@ -484,156 +258,10 @@ impl BacktestCollector {
         }
     }
 
-    /// Append a K-line record to CSV
-    async fn append_kline(
-        output_dir: &Path,
-        symbol: &str,
-        kline: &super::binance_klines::Kline,
-    ) -> Result<()> {
-        let path = output_dir.join("klines.csv");
-        let file = OpenOptions::new()
-            .append(true)
-            .open(&path)
-            .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        let mut writer = BufWriter::new(file);
-        writeln!(
-            writer,
-            "{},{},{},{},{},{},{},{},{}",
-            kline.open_time.timestamp(),
-            kline.open_time.format("%Y-%m-%d %H:%M:%S"),
-            symbol,
-            kline.open,
-            kline.high,
-            kline.low,
-            kline.close,
-            kline.volume,
-            kline.trades,
-        )
-        .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        Ok(())
-    }
-
-    /// Append a PM price record to CSV
-    async fn append_pm_price(
-        output_dir: &Path,
-        market: &ActiveMarket,
-        spot_price: Decimal,
-    ) -> Result<()> {
-        let path = output_dir.join("pm_prices.csv");
-        let file = OpenOptions::new()
-            .append(true)
-            .open(&path)
-            .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        let now = Utc::now();
-        let time_remaining = (market.resolution_time - now).num_seconds().max(0);
-
-        let mut writer = BufWriter::new(file);
-        writeln!(
-            writer,
-            "{},{},{},{},{},{},{},{},{},{},{},{},{},{},{},{}",
-            now.timestamp(),
-            now.format("%Y-%m-%d %H:%M:%S"),
-            market.market_id,
-            market.condition_id,
-            market.symbol,
-            market.threshold,
-            spot_price,
-            market.yes_price,
-            market.no_price,
-            market.yes_bid,
-            market.yes_ask,
-            market.no_bid,
-            market.no_ask,
-            market.resolution_time.timestamp(),
-            time_remaining,
-            "pending",
-        )
-        .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        Ok(())
-    }
-
-    /// Fetch active 15-minute markets for a symbol
-    async fn fetch_active_markets(
-        _client: &PolymarketClient,
-        symbol: &str,
-    ) -> Result<Vec<ActiveMarket>> {
-        // Map symbol to coin name
-        let coin = match symbol {
-            "BTCUSDT" => "BTC",
-            "ETHUSDT" => "ETH",
-            "SOLUSDT" => "SOL",
-            "XRPUSDT" => "XRP",
-            _ => return Ok(Vec::new()),
-        };
-
-        // Search for active 15-minute markets
-        // This is a simplified version - in production, you'd use the actual PM API
-        let _search_term = format!("{} 15", coin);
-
-        // Use client to search markets
-        // For now, return empty - this needs to be connected to actual PM API
-        // In the real implementation, this would call client.search_markets()
-
-        Ok(Vec::new())
-    }
-
-    /// Check if a market has resolved
-    async fn check_resolution(_client: &PolymarketClient, _condition_id: &str) -> Result<bool> {
-        // Check market resolution status
-        // This needs to be connected to actual PM API
-        // Returns true for YES, false for NO
-        Err(PloyError::Internal("Not implemented".to_string()))
-    }
-
-    /// Update outcome in CSV file
-    async fn update_outcome(output_dir: &Path, market_id: &str, outcome: bool) -> Result<()> {
-        // This is a simplified version - in production you'd update the CSV properly
-        // For now, we'll append a new line indicating resolution
-        let path = output_dir.join("resolutions.csv");
-
-        let file = OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(&path)
-            .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        let mut writer = BufWriter::new(file);
-        writeln!(
-            writer,
-            "{},{},{}",
-            Utc::now().timestamp(),
-            market_id,
-            if outcome { "YES" } else { "NO" },
-        )
-        .map_err(|e| PloyError::Internal(e.to_string()))?;
-
-        Ok(())
-    }
-
     /// Get collection statistics
     pub async fn stats(&self) -> CollectorStats {
         self.stats.read().await.clone()
     }
-}
-
-/// Active market information
-#[derive(Debug, Clone)]
-pub struct ActiveMarket {
-    pub market_id: String,
-    pub condition_id: String,
-    pub symbol: String,
-    pub threshold: Decimal,
-    pub yes_price: Decimal,
-    pub no_price: Decimal,
-    pub yes_bid: Decimal,
-    pub yes_ask: Decimal,
-    pub no_bid: Decimal,
-    pub no_ask: Decimal,
-    pub resolution_time: DateTime<Utc>,
 }
 
 // ============================================================================
